Crypto Project/requirements.txt
prophet
pandas
psycopg2-binary
sqlalchemy
python-dotenv